Implement product linking and product dependency in tabs
FossilOrigin-Name: 378e2d0eee82aba295f1f9240148a83cfdc9614f

+++ caisse/lib/Entities/Tab.php
@@ -89,14 +89,23 @@ class Tab extends Entity
 			throw new UserException('Cette note est close, impossible de modifier la note.');
 		}
 
-		$db = DB::getInstance();
-		$product = $db->first(POS::sql('SELECT p.*, c.name AS category_name, c.account AS category_account
-			FROM @PREFIX_products p
-			INNER JOIN @PREFIX_categories c ON c.id = p.category
-			WHERE p.id = ?'), $id);
+		$product = Products::get($id);
 
 		if (!$product) {
 			throw new UserException('This product does not exist: ' . $id);
+		}
+
+		$parent = $this->addProductItem($product, $user_weight, $price, $type);
+
+		foreach ($product->listLinkedProducts() as $p) {
+			$this->addProductItem($p, null, null, TabItem::TYPE_PRODUCT, $parent->id());
+		}
+	}
+
+	public function addProductItem(Product $product, ?string $user_weight = null, ?int $price = null, int $type = TabItem::TYPE_PRODUCT, ?int $id_parent = null): TabItem
+	{
+		if ($this->closed) {
+			throw new UserException('Cette note est close, impossible de modifier la note.');
 		}
 
 		$weight = $product->weight;
@@ -113,20 +122,23 @@ class Tab extends Entity
 
 		$item = new TabItem;
 		$item->importForm([
-			'tab'           => $this->id,
-			'product'       => (int)$product->id,
-			'qty'           => (int)$product->qty,
-			'price'         => $price,
-			'weight'        => $weight,
-			'name'          => $product->name,
-			'category_name' => $product->category_name,
-			'description'   => $product->description,
-			'account'       => $product->category_account,
-			'type'          => $type,
-			'pricing'       => $pricing,
-			'id_fee'        => $product->id_fee,
+			'tab'            => $this->id,
+			'product'        => (int)$product->id,
+			'qty'            => (int)$product->qty,
+			'price'          => $price,
+			'weight'         => $weight,
+			'name'           => $product->name,
+			'category_name'  => $product->category()->name,
+			'description'    => $product->description,
+			'account'        => $product->category()->account,
+			'type'           => $type,
+			'pricing'        => $pricing,
+			'id_fee'         => $product->id_fee,
+			'id_parent_item' => $id_parent,
 		]);
+
 		$item->save();
+		return $item;
 	}
 
 	public function getItem(int $id): ?TabItem
@@ -144,6 +156,10 @@ class Tab extends Entity
 
 		if (!$item) {
 			return;
+		}
+
+		if ($item->id_parent_item) {
+			throw new UserException('Ce produit est lié à un autre produit, il ne peut être supprimé seul, il faut supprimer le produit "parent".');
 		}
 
 		$item->delete();
@@ -316,6 +332,10 @@ class Tab extends Entity
 	}
 
 	public function rename(string $new_name, ?int $user_id) {
+		if ($this->closed) {
+			throw new UserException('Cette note est close, impossible de modifier la note.');
+		}
+
 		$new_name = trim($new_name);
 		$db = DB::getInstance();
 		return $db->update(POS::tbl('tabs'), ['name' => $new_name, 'user_id' => $user_id], $db->where('id', $this->id));
